Update Ingredient -----------------
- Add explicit toString() methods to entities to avoid the cyclic references in those generated by Lombok
- Delete ingredient detail page template as I intend to re-use the form or recipe edit page for this
- Turn the Mapstruct Mappers into singleton-scoped Spring beans which can be retrieved via @Autowired via @Mapper(componentMpdel = "spring"). This fits better into our Spring project
- Fixed the MapStruct Mapper classes properly to avoid cyclic references and StackOverflowError

!!!
PLEASE NOTE: In case of bi-directional relationships such mappings can trigger a StackOverflowError due to circular references.
MapStruct can handle primitive and standard classes like String and Integer just fine, but if you are using custom complex objects it needs some instructions on how it should map these for example:
		@Mapper(uses = {BookMapper.class, IngredientMapper.class})
		public interface RecipeMapper { }

If you would stop there and try to map the classes this way, you will get hit by the StackOverflowError due to circular references you have now defined (Recipe contains ingredients which has a property recipe which has ingredients...).
Such default Mapper setups can only be used if there is no bidirectional relationship which would trigger the inverse mapping as well.

	You could write this down like A -> B -> A -> B -> A ...

Concerning the object mapping, my experience has shown that you should be able to map this out like: A -> B -> A (excluding the relations this time to break the cycle)	for both Entity to DTO and DTO to Entity mappings. This enables you to:
* Drill down to associated objects in the frontend: ex. display a list of ingredients for a recipe
* Persist the inverse relation when saving an object: ex. If you would only map A -> B. The IngredientDTOs inside RecipeDTO would not have a recipe property and when saving an ingredient you would need to pass in the recipe id as a parameter and jump through some hoops to associate the ingredient entity object with a recipe entity object before saving the ingredient entity to the database.

Defining the mappings like A -> B -> A (excluding the relations this time to break the cycle) will boil down to defining separate mappings for when you want to exclude the related complex objects from the mapping at the point where you want to break the cycle.

	@IterableMapping(qualifiedByName = "<MAPPING_NAME>")	is used to map a collection of complex objects, which refers to a mapping for a single complex object.

	@Mapping(target = "PropertyName", qualifiedByName = "<MAPPING_NAME>")				can be used to point to an alternative mapping which excludes the inverse relationships when mapping a collection of complex objects (when you want to break the cycle)

	@Mapping(target = "<PropertyName>[.<childPropertyName>]", ignore = true)			can be used to indicate that a property of an object should not be mapped at all. So this can be used to completely leave out a (collection of) complex object(s) entirely or to ignore properties inside of a single (not a collection) related complex objects directly in case they are not needed.

If you don't use the qualifiedByName attribute and the matching @Named() annotations, your mappings will not compile with an error about Ambiguous mappings if you have multiple methods with the same return type and input parameter types in the Mapper interface.

It may be a good practice use method names matching the @Named annotation value in case you use named mappings.
!!!